+++ tests/Application/Query/Agenda/AgendaViewQueryHandlerTest.php
@@ -101,7 +101,7 @@ class AgendaViewQueryHandlerTest extends TestCase
         $sheetRepository->isUserParticipantMultipleSheetsInEvent($user, $event)->shouldBeCalled()->willReturn(true);
 
         $happeningParticipationRepository = $this->prophesize(HappeningParticipationRepositoryInterface::class);
-        $happeningParticipationRepository->findByUser($user, $event, true)->shouldBeCalled()->willReturn([
+        $happeningParticipationRepository->findByUser($user, $event, true, true)->shouldBeCalled()->willReturn([
             $happeningParticipation,
         ]);
         $happeningParticipationRepository->findBySpeaker($user, $event)->shouldBeCalled()->willReturn([]);
@@ -247,7 +247,7 @@ class AgendaViewQueryHandlerTest extends TestCase
 
         $happeningParticipationRepository = $this->prophesize(HappeningParticipationRepositoryInterface::class);
         $happeningParticipationRepository
-            ->findByUser($user2, $event, true)
+            ->findByUser($user2, $event, true, true)
             ->shouldBeCalled()
             ->willReturn([
                 $happeningParticipation1,